refactor(@packages/db-drizzlepg): some code cleanup
* surrogate key helper
* rename inferred row types
* extract query logger to own file
